requirements.txt
streamlit
openai
python-dotenv
pydantic
nest_asyncio
